refactor: 중복된 페이지 정보 클래스를 제거하고 하나로 통일한다.
