+++ src/main/java/com/clova/anifriends/domain/recruitment/dto/response/FindShelterRecruitmentsResponse.java
@@ -1,6 +1,6 @@
 package com.clova.anifriends.domain.recruitment.dto.response;
 
-import com.clova.anifriends.domain.common.dto.PageInfo;
+import com.clova.anifriends.domain.common.PageInfo;
 import com.clova.anifriends.domain.recruitment.Recruitment;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -33,7 +33,7 @@ public record FindShelterRecruitmentsResponse(
 
     public static FindShelterRecruitmentsResponse from(Page<Recruitment> recruitments) {
         return new FindShelterRecruitmentsResponse(
-            PageInfo.from(recruitments),
+            PageInfo.of(recruitments.getTotalElements(), recruitments.hasNext()),
             recruitments.stream()
                 .map(RecruitmentResponse::from)
                 .toList()
